requirements-locked.txt
alembic==1.17.0
black==25.9.0
blinker==1.9.0
cfgv==3.4.0
click==8.3.0
distlib==0.4.0
filelock==3.20.0
flake8==7.3.0
Flask==3.1.2
Flask-Admin==2.0.0
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
gitdb==4.0.12
GitPython==3.1.41
greenlet==3.2.4
identify==2.6.15
iniconfig==2.3.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.3.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
python-dotenv==1.2.1
pytokens==0.2.0
PyYAML==6.0.3
setuptools==78.1.1
smmap==5.0.2
SQLAlchemy==2.0.44
typing_extensions==4.15.0
virtualenv==20.35.3
Werkzeug==3.1.3
WTForms==3.2.1
